reoganize logic tree modules to make them smaller
