+++ docs/api/logic_tree/branch.md
@@ -1,0 +1,3 @@
+::: nzshm_model.logic_tree.branch
+    options:
+        filters: ["!^_"]
