+++ docs/api/logic_tree/correlation.md
@@ -1,0 +1,3 @@
+::: nzshm_model.logic_tree.correlation
+    options:
+        filters: ["!^_"]
