+++ nzshm_model/logic_tree/logic_tree_base.py
@@ -12,12 +12,14 @@ from functools import reduce
 from itertools import product
 from operator import mul
 from pathlib import Path
-from typing import Any, Dict, Generator, Iterator, List, Optional, Type, Union, overload
+from typing import Any, Dict, Generator, Iterator, List, Type, Union
 
 import dacite
 
 from nzshm_model.psha_adapter import PshaAdapterInterface
 
+from .branch import Branch, CompositeBranch
+from .correlation import LogicTreeCorrelations
 
 # TODO:
 # - move values to the base class?
@@ -25,32 +27,6 @@ from nzshm_model.psha_adapter import PshaAdapterInterface
 #    the initializer of FilteredBranch?
 # - FilteredBranch doesn't need to be a data class as it should not be serialized and doesn't contain many arguments
 # - should we use FilteredBranch for correlation so the branches can be traced back to the BranchSet?
-@dataclass
-class Branch(ABC):
-    """
-    Abstract baseclass for logic tree branches
-
-    Arguments:
-        name: a name for the branch
-        weight: a weight for the branch
-    """
-
-    name: str = ""
-    weight: float = 1.0
-
-    @abstractmethod
-    def filtered_branch(self, logic_tree: 'LogicTree', branch_set: 'BranchSet') -> 'FilteredBranch':
-        """
-        Produce a new filtered branch with the properties of the branch
-
-        Parameters:
-            logic_tree: The logic tree that the branch belongs to
-            branch_set: The branch est that the branch belongs to
-
-        Returns:
-            a filtered branch
-        """
-        pass
 
 
 @dataclass
@@ -83,119 +59,6 @@ class BranchSet:
         for b in self.branches:
             weight += b.weight
         return math.isclose(weight, 1.0)
-
-
-@dataclass
-class Correlation:
-    """
-    A correlation between **BranchSet**s of a logic tree. Correlations are used when buiding combinations of branches
-    from differnt **BranchSet**s.
-
-    For example, if a logic tree contains branchets A and B each with branches A1, A2 and B1, B2, respectivly, then
-    a correlation A1, B1 will only allow A1-B1 not A1-B2 as valid composite branches.
-
-    The primary_brach will not appear in combination with any branches except those identified in associated_branches
-    for the relevent **BranchSet**s
-
-    Arguments:
-        primary_branch: the branch that **MUST** be correlated with the associated brances.
-        associated_branches: list of branches that the primary_branch must always be coupled with.
-        weight: weight used for composite branch formed by correlation. Defaults to weight of primary_branch
-    """
-
-    primary_branch: Branch = field(default_factory=Branch)
-    associated_branches: List[Branch] = field(default_factory=list)
-    weight: Optional[float] = None
-
-    def __post_init__(self):
-        self.weight = self.primary_branch.weight if not self.weight else self.weight
-
-    @property
-    def all_branches(self) -> List[Branch]:
-        """
-        list of all branches in a correlation; both the primary branch and the assiciated branches
-
-        Returns:
-            list of branches in correlation
-        """
-        return [self.primary_branch] + self.associated_branches
-
-
-@dataclass(frozen=True)
-class LogicTreeCorrelations(Sequence):
-    """
-    All correlations for a logic tree.
-
-    Arguments:
-        correlation_groups: list of correlations to be applied to the logic tree branch sets.
-    """
-
-    correlation_groups: List[Correlation] = field(default_factory=list)
-
-    def __post_init__(self) -> None:
-        self._validate_correlations()
-
-    def _validate_correlations(self) -> None:
-        """
-        check that there are no repeats in the 0th element of each correlation
-        """
-        prim_branches = list(self.primary_branches())
-        if len([branch for branch in prim_branches if prim_branches.count(branch) > 1]) != 0:
-            raise ValueError("there is a repeated branch in the 0th element of the correlations")
-
-    def primary_branches(self) -> Generator[Branch, None, None]:
-        """
-        Yield the primary branches of all correlation_groups
-
-        Returns:
-            branches
-        """
-        for cor in self.correlation_groups:
-            yield cor.primary_branch
-
-    @overload
-    def __getitem__(self, i: int) -> Correlation:
-        ...
-
-    @overload
-    def __getitem__(self, i: slice) -> Sequence[Correlation]:
-        ...
-
-    def __getitem__(self, i):
-        if isinstance(i, slice):
-            raise TypeError("LogicTreeCorrelations does not support slicing")
-        return self.correlation_groups[i]
-
-    def __len__(self) -> int:
-        return len(self.correlation_groups)
-
-
-@dataclass
-class CompositeBranch:
-    """
-    A logic tree branch comprised of combinations of branches from one or more branch sets.
-
-    Arguments:
-        branches: the component branches (branches from branch sets) in the composite branch
-        weight: the weight of the composite branch
-    """
-
-    branches: Sequence[Branch] = field(default_factory=list)
-    weight: float = 1.0
-
-    def __post_init__(self) -> None:
-        self.weight = reduce(mul, [branch.weight for branch in self.branches], 1.0)
-
-    def __iter__(self):
-        self.__counter = 0
-        return self
-
-    def __next__(self):
-        if self.__counter >= len(self.branches):
-            raise StopIteration
-        else:
-            self.__counter += 1
-            return self.branches[self.__counter - 1]
 
 
 @dataclass
